requirements.txt
asgiref==3.8.1
Django==3.2.3
django-model-utils==4.5.0
djangorestframework==3.15.1
flake8==7.0.0
isort==5.13.2
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.11.0
